scripts: make ctrl & alt + click work as expected on site links
I noticed that our sidebar and table-of-content links overrode click
interactions (to allow soft navigation) even when the control or alt
keys are maintained which are often relied on respectively to open a
link in a new tab or a new window.

I suppose those shortcuts depend on the browser but I thought that as a
quick fix, it made sense to not actually do soft navigation (on the same
page) when one of those two keys is maintained during a click.